# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sports-Point API",
    "version": "1.0.0",
    "description": "运动会积分与成绩管理系统对外 API 文档。"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/v1/settings/report-environment": {
      "post": {
        "tags": [
          "公示"
        ],
        "summary": "保存公示环境信息",
        "description": "保存成绩公示单使用的日期、天气、风向、温度等环境信息。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportEnvironmentRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/ReportEnvironmentRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/notices/personal-result.xlsx": {
      "get": {
        "tags": [
          "公示"
        ],
        "summary": "导出个人成绩公示 XLSX",
        "description": "按项目和模板导出个人项目成绩公示单。",
        "responses": {
          "200": {
            "description": "XLSX 文件",
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "event_id",
            "in": "query",
            "required": true,
            "description": "项目 ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "template_name",
            "in": "query",
            "required": true,
            "description": "XLSX 模板文件名",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/notices/personal-result.pdf": {
      "get": {
        "tags": [
          "公示"
        ],
        "summary": "预览个人成绩公示 PDF",
        "description": "按项目和模板生成个人项目成绩公示 PDF。",
        "responses": {
          "200": {
            "description": "PDF 文件",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "event_id",
            "in": "query",
            "required": true,
            "description": "项目 ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "template_name",
            "in": "query",
            "required": true,
            "description": "XLSX 模板文件名",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/notices/team-result.xlsx": {
      "get": {
        "tags": [
          "公示"
        ],
        "summary": "导出团体成绩公示 XLSX",
        "description": "按项目和模板导出团体项目成绩公示单。",
        "responses": {
          "200": {
            "description": "XLSX 文件",
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "event_id",
            "in": "query",
            "required": true,
            "description": "项目 ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "template_name",
            "in": "query",
            "required": true,
            "description": "XLSX 模板文件名",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/notices/team-result.pdf": {
      "get": {
        "tags": [
          "公示"
        ],
        "summary": "预览团体成绩公示 PDF",
        "description": "按项目和模板生成团体项目成绩公示 PDF。",
        "responses": {
          "200": {
            "description": "PDF 文件",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "event_id",
            "in": "query",
            "required": true,
            "description": "项目 ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "template_name",
            "in": "query",
            "required": true,
            "description": "XLSX 模板文件名",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/notices/personal-attempt.xlsx": {
      "get": {
        "tags": [
          "公示"
        ],
        "summary": "导出个人检录表 XLSX",
        "description": "按项目和模板导出个人项目检录表（含每次尝试记录及作废标记）。",
        "responses": {
          "200": {
            "description": "XLSX 文件",
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "event_id",
            "in": "query",
            "required": true,
            "description": "项目 ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "template_name",
            "in": "query",
            "required": true,
            "description": "XLSX 模板文件名",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/notices/personal-attempt.pdf": {
      "get": {
        "tags": [
          "公示"
        ],
        "summary": "预览个人检录表 PDF",
        "description": "按项目和模板生成个人项目检录表 PDF。",
        "responses": {
          "200": {
            "description": "PDF 文件",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "event_id",
            "in": "query",
            "required": true,
            "description": "项目 ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "template_name",
            "in": "query",
            "required": true,
            "description": "XLSX 模板文件名",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/notices/team-attempt.xlsx": {
      "get": {
        "tags": [
          "公示"
        ],
        "summary": "导出团体检录表 XLSX",
        "description": "按项目和模板导出团体项目检录表（含每次尝试记录及作废标记）。",
        "responses": {
          "200": {
            "description": "XLSX 文件",
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "event_id",
            "in": "query",
            "required": true,
            "description": "项目 ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "template_name",
            "in": "query",
            "required": true,
            "description": "XLSX 模板文件名",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/notices/team-attempt.pdf": {
      "get": {
        "tags": [
          "公示"
        ],
        "summary": "预览团体检录表 PDF",
        "description": "按项目和模板生成团体项目检录表 PDF。",
        "responses": {
          "200": {
            "description": "PDF 文件",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "event_id",
            "in": "query",
            "required": true,
            "description": "项目 ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "template_name",
            "in": "query",
            "required": true,
            "description": "XLSX 模板文件名",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "OkResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "ok"
        ],
        "additionalProperties": true
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "参数错误"
          }
        },
        "required": [
          "ok",
          "error"
        ]
      },
...
      "ReportEnvironmentRequest": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "日期"
          },
          "wind_direction": {
            "type": "string",
            "description": "风向"
          },
          "wind_speed": {
            "type": "string",
            "description": "风速"
          },
          "air_quality": {
            "type": "string",
            "description": "空气质量"
          },
          "weather": {
            "type": "string",
            "description": "天气"
          },
          "temperature_high": {
            "type": "string",
            "description": "最高温"
          },
          "temperature_low": {
            "type": "string",
            "description": "最低温"
          }
        }
      },
...
    }
  }
}